requirements.txt
torch==1.8.1
torchvision==0.9.1
torchaudio==0.8.1
numpy==1.19.5
Pillow==7.0.0
opencv-python==4.5.2.52
natsort